[CHORE] ChangeViewController 명칭 변경 및 디렉토리 추가(#116)
