+++ EarthValley80/EarthValley80/Screen/SideTabbar/SideTabbarViewController.swift
@@ -9,41 +9,41 @@ import Combine
 import SwiftUI
 import UIKit
 
-final class ChangeViewController: UIViewController {
-    // MARK: - property
-    let embeddedYomoJomoNewsViewController = YomojomoNewsViewController()
-    let embeddedMyNewsDrawerViewController = MyNewsDrawerViewController()
-    private lazy var notifier = EventManager.shared
-    private var cancellables = Set<AnyCancellable>()
-    
-    // MARK: - life cycle
-    override func viewDidLayoutSubviews() {
-        super.viewDidLayoutSubviews()
-    }
-    
-    override func viewDidLoad() {
-        view.addSubview(self.embeddedYomoJomoNewsViewController.view)
-        view.addSubview(self.embeddedMyNewsDrawerViewController.view)
-        
-        notifier
-            .sideMenuNumber
-            .sink { [weak self] currentIndex in
-                switch currentIndex {
-                case 1:
-                    self?.embeddedYomoJomoNewsViewController.view.isHidden = false
-                    self?.embeddedMyNewsDrawerViewController.view.isHidden = true
-                case 2:
-                    self?.embeddedYomoJomoNewsViewController.view.isHidden = true
-                    self?.embeddedMyNewsDrawerViewController.view.isHidden = true
-                case 3:
-                    self?.embeddedYomoJomoNewsViewController.view.isHidden = true
-                    self?.embeddedMyNewsDrawerViewController.view.isHidden = false
-                default: break
-                }
-            }
-            .store(in: &cancellables)
-    }
-}
+//final class ChangeViewController: UIViewController {
+//    // MARK: - property
+//    let embeddedYomoJomoNewsViewController = YomojomoNewsViewController()
+//    let embeddedMyNewsDrawerViewController = MyNewsDrawerViewController()
+//    private lazy var notifier = EventManager.shared
+//    private var cancellables = Set<AnyCancellable>()
+//
+//    // MARK: - life cycle
+//    override func viewDidLayoutSubviews() {
+//        super.viewDidLayoutSubviews()
+//    }
+//
+//    override func viewDidLoad() {
+//        view.addSubview(self.embeddedYomoJomoNewsViewController.view)
+//        view.addSubview(self.embeddedMyNewsDrawerViewController.view)
+//
+//        notifier
+//            .sideMenuNumber
+//            .sink { [weak self] currentIndex in
+//                switch currentIndex {
+//                case 1:
+//                    self?.embeddedYomoJomoNewsViewController.view.isHidden = false
+//                    self?.embeddedMyNewsDrawerViewController.view.isHidden = true
+//                case 2:
+//                    self?.embeddedYomoJomoNewsViewController.view.isHidden = true
+//                    self?.embeddedMyNewsDrawerViewController.view.isHidden = true
+//                case 3:
+//                    self?.embeddedYomoJomoNewsViewController.view.isHidden = true
+//                    self?.embeddedMyNewsDrawerViewController.view.isHidden = false
+//                default: break
+//                }
+//            }
+//            .store(in: &cancellables)
+//    }
+//}
 
 final class SideTabbarViewController: UIViewController {
     // MARK: - property
@@ -56,7 +56,7 @@ final class SideTabbarViewController: UIViewController {
     private let embeddedYomoJomoNewsViewController = YomojomoNewsViewController()
     private let embeddedMyNewsDrawerViewController = MyNewsDrawerViewController()
     private var cancellables = Set<AnyCancellable>()
-    private let changeVC = ChangeViewController()
+    private let changeVC = ContainerViewController()
     
     // MARK: - life cycle
     override func viewDidLoad() {
